Wire up logic for determining numbr of records sent

+++ p2p-lib/src/main/java/org/smartregister/p2p/search/ui/P2PReceiverViewModel.kt
@@ -228,7 +228,7 @@ class P2PReceiverViewModel(
     return sendingDeviceAppLifetimeKey
   }
 
-  override fun updateProgress(resStringMsg: Int, recordSize: Int) {
+  override fun updateProgress(resStringMsg: Int, recordSize: Long) {
     // TODO Update UI with record size
   }
